+++ platform/model_get_image_response.go
@@ -25,5 +25,5 @@ type GetImageResponse struct {
 	Errors []ResponseError `json:"errors,omitempty"`
 	// The operation time in microseconds.  This is the time it took to process
 	// the request and generate the response.
-	OpTimeUs *int32 `json:"op_time_us,omitempty"`
+	OpTimeUs *uint64 `json:"op_time_us,omitempty"`
 }
