+++ platform/model_get_instance_logs_response_logged_instance_range.go
@@ -11,7 +11,7 @@ package platform
 
 type GetInstanceLogsResponseLoggedInstanceRange struct {
 	// The first retrieved byte.
-	Start *int32 `json:"start,omitempty"`
+	Start *uint64 `json:"start,omitempty"`
 	// The last retrieved byte.
-	End *int32 `json:"end,omitempty"`
+	End *uint64 `json:"end,omitempty"`
 }
