+++ platform/model_quotas_response.go
@@ -24,7 +24,7 @@ type QuotasResponse struct {
 	Errors []ResponseError `json:"errors,omitempty"`
 	// The operation time in microseconds.  This is the time it took to process
 	// the request and generate the response.
-	OpTimeUs *int32 `json:"op_time_us,omitempty"`
+	OpTimeUs *uint64 `json:"op_time_us,omitempty"`
 	// An optional message providing additional information about the response.
 	Message *string `json:"message,omitempty"`
 }
