+++ platform/model_response_error.go
@@ -10,5 +10,5 @@ package platform
 
 type ResponseError struct {
 	// The HTTP status code of the error.
-	Status *int32 `json:"status,omitempty"`
+	Status *uint64 `json:"status,omitempty"`
 }
